feat: Implement DateItem composable
* feat: Implement CategoryItem component

This commit introduces the `CategoryItem` composable, a reusable UI
component for displaying category information. It includes an icon,
title, and optional count and selection indicator.  It also adds
the necessary vector drawables for the category icons.

* refactor: Reduce size of category item title

Updated the typography style for the category item title from
`labelMedium` to `labelSmall` to better fit within the item's
layout and improve visual balance.

* feat: Implement DateItem composable

This commit introduces the DateItem composable, a UI component
that displays a date and day of the week. It includes styling
for selected and unselected states, with different background
gradients and text colors. The component is also clickable.
A preview is included for easy testing and visualization.

* fix: Adjust check icon color in CategoryItem

The check icon color in the CategoryItem was previously hardcoded
to white. This change updates the color to use the `onPrimary`
color from the `TudeeTheme` with an alpha of 0.87 for better
contrast and consistency with the application's theme.

* refactor: Use TudeeTheme colors for DateItem gradient

Replaces hardcoded colors in the DateItem composable's
gradient with colors from the TudeeTheme. This promotes
consistency and maintainability across the application's
UI.

* refactor: Use TudeeTheme shapes and spacing in DateItem

Use predefined theme values for shape and padding in DateItem
for consistency.

* feat: Improve DateItem text visibility

Updated DateItem's day of week text style to `labelMedium`
to improve readability. Also, removed unused `app_name` string
resource.

* feat: Refactor DateItem and theme colors

Updates DateItem background and adjusts light and dark theme colors
for improved visual consistency and aesthetics.

* refactor: Adjust color values for light and dark themes

Refactors color definitions in LightThemeColors.kt and
DarkThemeColors.kt to improve visual consistency and contrast.
This includes changes to surface colors, text colors, and
various accent colors for better readability and aesthetics.

* refactor: Rename `day` to `dayOfMonth` in DateItem

This change renames the `day` property in the `DateItem`
composable to `dayOfMonth` for clarity and consistency.