feat: Implement CategoryItem component
* feat: Implement CategoryItem component

This commit introduces the `CategoryItem` composable, a reusable UI
component for displaying category information. It includes an icon,
title, and optional count and selection indicator.  It also adds
the necessary vector drawables for the category icons.

* refactor: Reduce size of category item title

Updated the typography style for the category item title from
`labelMedium` to `labelSmall` to better fit within the item's
layout and improve visual balance.

* fix: Improve category item check icon color

Updates the color of the check icon in the CategoryItem
component to use the onPrimary color from the TudeeTheme with
an alpha of 0.87. This improves visibility and consistency
with the application's theme.

* refactor: Increase CategoryItem icon and badge size

Increase the size of the icon and badge in the
CategoryItem composable for better visibility and
improved user experience. Also slightly adjusted offset and width.

* feat: Change CategoryItem background color

Update background color of CategoryItem to surfaceHigh for
better visual hierarchy.